[mod] [filemanager] [toolbox]  5.0外置sd卡问题

+++ FileManager/src/com/filemanager/dialogs/RenameDialog.java
@@ -1,10 +1,14 @@
 package com.filemanager.dialogs;
 
 import android.app.Dialog;
+import android.content.Context;
 import android.os.Bundle;
 import android.support.v4.app.DialogFragment;
+import android.support.v4.provider.DocumentFile;
 import android.text.InputType;
+import android.text.TextUtils;
 import android.widget.Toast;
+import base.util.FileUtils;
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.filemanager.R;
 import com.filemanager.files.FileHolder;
@@ -16,11 +20,12 @@ import java.io.File;
 
 public class RenameDialog extends DialogFragment {
 	private FileHolder mFileHolder;
+    private Context mContext;
 	
 	@Override
 	public void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-		
+        mContext = getActivity().getApplicationContext();
 		mFileHolder = getArguments().getParcelable(FileManagerIntents.EXTRA_DIALOG_FILE_HOLDER);
 	}
 
@@ -40,23 +45,40 @@ public class RenameDialog extends DialogFragment {
                         }).build();
     }
 	
-	private void renameTo(String to){
-		boolean res = false;
-		
-		if(to.length() > 0){
-			File from = mFileHolder.getFile();
-			
-			File dest = new File(mFileHolder.getFile().getParent() + File.separator + to);
-			if(!dest.exists()){
-				res = mFileHolder.getFile().renameTo(dest);
-				((FileListFragment) getTargetFragment()).refresh();
+	private void renameTo(String to) {
+        boolean res = false;
 
-				// Inform media scanner
-				MediaScannerUtils.informFileDeleted(getActivity().getApplicationContext(), from);
-				MediaScannerUtils.informFileAdded(getActivity().getApplicationContext(), dest);
-			}
-		}
-		
-		Toast.makeText(getActivity(), res ? R.string.rename_success : R.string.rename_failure, Toast.LENGTH_SHORT).show();
-	}
+        if (to.length() > 0) {
+            File from = mFileHolder.getFile();
+            File dest = new File(mFileHolder.getFile().getParent() + File.separator + to);
+
+            if (FileUtils.isAndroid5() && FileUtils.isOnExtSdCard(from, mContext) && !dest.exists() && !TextUtils.isEmpty(to)) {
+                DocumentFile documentFile;
+                if (from.isDirectory()) {
+                    documentFile = FileUtils.getDocumentFile(from, true, false, mContext);
+                } else {
+                    documentFile = FileUtils.getDocumentFile(from, false, false, mContext);
+                }
+                if(documentFile != null) {
+                    res = documentFile.renameTo(to);
+                    ((FileListFragment) getTargetFragment()).refresh();
+                    // Inform media scanner
+                    MediaScannerUtils.informFileDeleted(getActivity().getApplicationContext(), from);
+                    MediaScannerUtils.informFileAdded(getActivity().getApplicationContext(), dest);
+                }
+
+            } else {
+                if (!dest.exists()) {
+                    res = mFileHolder.getFile().renameTo(dest);
+                    ((FileListFragment) getTargetFragment()).refresh();
+                    // Inform media scanner
+                    MediaScannerUtils.informFileDeleted(getActivity().getApplicationContext(), from);
+                    MediaScannerUtils.informFileAdded(getActivity().getApplicationContext(), dest);
+                }
+            }
+
+        }
+
+        Toast.makeText(getActivity(), res ? R.string.rename_success : R.string.rename_failure, Toast.LENGTH_SHORT).show();
+    }
 }
